// packages/moim-web/app/actions/types.tsx
export enum GroupTypes {
  SUCCEEDED_GET_INITIAL_GROUP_ID = "GROUP.SUCCEEDED_GET_INITIAL_GROUP_ID",

  START_BOOT_STRAP = "GROUP.START_BOOT_STRAP",
  SUCCESS_BOOT_STRAP = "GROUP.SUCCESS_BOOT_STRAP",
  FAILED_BOOT_STRAP = "GROUP.FAILED_BOOT_STRAP",

  START_VALIDATE_GROUP_DOMAIN = "GROUP.START_VALIDATE_GROUP_DOMAIN",
  VALIDATE_GROUP_DOMAIN = "GROUP.VALIDATE_GROUP_DOMAIN",

  START_FETCHING_LIST_GROUP = "GROUP.START_FETCHING_LIST_GROUP",
  FETCHED_LIST_GROUP = "GROUP.FETCHED_LIST_GROUP",
  FAILED_FETCHING_LIST_GROUP = "GROUP.FAILED_FETCHING_LIST_GROUP",

  START_UPDATE_GROUP_ICON = "GROUP.START_UPDATE_GROUP_ICON",
  SUCCEED_UPDATE_GROUP_ICON = "GROUP.SUCCEED_UPDATE_GROUP_ICON",
  FAILED_UPDATE_GROUP_ICON = "GROUP.FAILED_UPDATE_GROUP_ICON",

  START_CROP_GROUP_ICON = "GROUP.START_CROP_GROUP_ICON",
  SUCCEEDED_CROP_GROUP_ICON = "GROUP.SUCCEEDED_CROP_GROUP_ICON",
  FAILED_CROP_GROUP_ICON = "GROUP.FAILED_CROP_GROUP_ICON",

  START_CROP_GROUP_BANNER = "GROUP.START_CROP_GROUP_BANNER",
  SUCCEEDED_CROP_GROUP_BANNER = "GROUP.SUCCEEDED_CROP_GROUP_BANNER",
  FAILED_CROP_GROUP_BANNER = "GROUP.FAILED_CROP_GROUP_BANNER",

  START_UPDATE_GROUP_BANNER = "GROUP.START_UPDATE_GROUP_BANNER",
  SUCCEEDED_UPDATE_GROUP_BANNER = "GROUP.SUCCEEDED_UPDATE_GROUP_BANNER",
  FAILED_UPDATE_GROUP_BANNER = "GROUP.FAILED_UPDATE_GROUP_BANNER",

  START_CREATE_GROUP = "GROUP.START_CREATE_GROUP",
  SUCCEEDED_CREATE_GROUP = "GROUP.SUCCEEDED_CREATE_GROUP",
  FAILED_CREATE_GROUP = "GROUP.FAILED_CREATE_GROUP",

  START_CREATE_SUB_GROUP = "GROUP.START_CREATE_SUB_GROUP",
  SUCCEEDED_CREATE_SUB_GROUP = "GROUP.SUCCEEDED_CREATE_SUB_GROUP",
  FAILED_CREATE_SUB_GROUP = "GROUP.FAILED_CREATE_SUB_GROUP",

  START_GET_GROUP_DATA = "GROUP.START_GET_GROUP_DATA",
  SUCCEEDED_GET_GROUP_DATA = "GROUP.SUCCEEDED_GET_GROUP_DATA",
  FAILED_GET_GROUP_DATA = "GROUP.FAILED_GET_GROUP_DATA",

  START_FETCHING_GROUP_STATUS = "GROUP.START_FETCHING_GROUP_STATUS",
  FETCHED_GROUP_STATUS = "GROUP.FETCHED_GROUP_STATUS",
  FAILED_FETCHING_GROUP_STATUS = "GROUP.FAILED_FETCHING_GROUP_STATUS",

  START_FETCHING_SUB_MOIMS = "GROUP.START_FETCHING_SUB_MOIMS",
  FETCHED_SUB_MOIMS = "GROUP.FETCHED_SUB_MOIMS",
  FAILED_FETCHING_SUB_MOIMS = "GROUP.FAILED_FETCHING_SUB_MOIMS",

  START_GET_JOINED_SUB_MOIMS = "GROUP.START_GET_JOINED_SUB_MOIMS",
  SUCCEED_GET_JOINED_SUB_MOIMS = "GROUP.SUCCEED_FETCHED_JOINED_SUB_MOIMS",
  FAILED_GET_JOINED_SUB_MOIMS = "GROUP.FAILED_GET_JOINED_SUB_MOIMS",

  OPEN_JOIN_GROUP_DIALOG = "GROUP.OPEN_JOIN_GROUP_DIALOG",
  CLOSE_JOIN_GROUP_DIALOG = "GROUP.CLOSE_JOIN_GROUP_DIALOG",

  CLEAR_MOIM_ICON = "GROUP.CLEAR_MOIM_ICON",

  START_RENAME_MOIM = "GROUP.START_RENAME_MOIM",
  SUCCEED_RENAME_MOIM = "GROUP.SUCCEED_RENAME_MOIM",
  FAILED_RENAME_MOIM = "GROUP.FAILED_RENAME_MOIM",

  START_SET_DESCRIPTION_MOIM = "GROUP.START_SET_DESCRIPTION_MOIM",
  SUCCEED_SET_DESCRIPTION_MOIM = "GROUP.SUCCEED_SET_DESCRIPTION_MOIM",
  FAILED_SET_DESCRIPTION_MOIM = "GROUP.FAILED_SET_DESCRIPTION_MOIM",

  START_GET_MOIM_THEME = "GROUP.START_GET_MOIM_THEME",
  SUCCEED_GET_MOIM_THEME = "GROUP.SUCCEED_GET_MOIM_THEME",
  FAILED_GET_MOIM_THEME = "GROUP.FAILED_GET_MOIM_THEME",

  START_GET_MOIM_COVER = "GROUP.START_GET_MOIM_COVER",
  SUCCEED_GET_MOIM_COVER = "GROUP.SUCCEED_GET_MOIM_COVER",
  FAILED_GET_MOIM_COVER = "GROUP.FAILED_GET_MOIM_COVER",

  START_UPDATE_HOME_CHANNEL = "GROUP.START_UPDATE_DEFAULT_CHANNEL",
  SUCCEED_UPDATE_HOME_CHANNEL = "GROUP.SUCCEED_UPDATE_DEFAULT_CHANNEL",
  FAILED_UPDATE_HOME_CHANNEL = "GROUP.FAILED_UPDATE_DEFAULT_CHANNEL",

  START_GET_PARENT_THEME = "GROUP.START_GET_PARENT_THEME",
  SUCCEED_GET_PARENT_THEME = "GROUP.SUCCEED_GET_PARENT_THEME",
  FAILED_GET_PARENT_THEME = "GROUP.FAILED_GET_PARENT_THEME",

  START_GET_RECOMMEND_MOIMS = "GROUP.START_GET_RECOMMEND_MOIMS",
  SUCCEED_GET_RECOMMEND_MOIMS = "GROUP.SUCCEED_GET_RECOMMEND_MOIMS",
  FAILED_GET_RECOMMEND_MOIMS = "GROUP.FAILED_GET_RECOMMEND_MOIMS",

  START_SEARCH_MOIMS = "GROUP.START_SEARCH_MOIMS",
  SUCCEED_SEARCH_MOIMS = "GROUP.SUCCEED_SEARCH_MOIMS",
  FAILED_SEARCH_MOIMS = "GROUP.FAILED_SEARCH_MOIMS",
  CLEAR_SEARCH_MOIMS = "GROUP.CLEAR_SEARCH_MOIMS",

  START_GET_INSTALLED_PLUGINS = "GROUP.START_GET_INSTALLED_PLUGINS",
  SUCCEED_GET_INSTALLED_PLUGINS = "GROUP.SUCCEED_GET_INSTALLED_PLUGINS",
  FAILED_GET_INSTALLED_PLUGINS = "GROUP.FAILED_GET_INSTALLED_PLUGINS",
}
// ...
export enum UserTypes {
  START_SEARCH_BY_KEYWORD = "USER.START_SEARCH_BY_KEYWORD",
  SUCCEED_SEARCH_BY_KEYWORD = "USER.SUCCEED_SEARCH_BY_KEYWORD",
  FAILED_SEARCH_BY_KEYWORD = "USER.FAILED_SEARCH_BY_KEYWORD",

  START_GET_USERS = "USER.START_GET_USERS",
  SUCCEED_GET_USERS = "USER.SUCCEED_GET_USERS",
  FAILED_GET_USERS = "USER.FAILED_GET_USERS",

  CLEAR_SEARCHED_USERS = "USER.CLEAR_SEARCHED_USERS",

  START_GET_SEARCH_USERS = "USER.START_GET_SEARCH_USERS",
  SUCCEED_GET_SEARCH_USERS = "USER.SUCCEED_GET_SEARCH_USERS",
  FAILED_GET_SEARCH_USERS = "USER.FAILED_GET_SEARCH_USERS",

  START_POST_USER = "USER.START_POST_USER",
  SUCCEED_POST_USER = "USER.SUCCEED_POST_USER",
  FAILED_POST_USER = "USER.FAILED_POST_USER",

  START_GET_BATCH_USERS = "USER.START_GET_BATCH_USERS",
  SUCCEED_GET_BATCH_USERS = "USER.SUCCEED_GET_BATCH_USERS",
  FAILED_GET_BATCH_USERS = "USER.FAILED_GET_BATCH_USERS",

  START_GET_PARENT_MOIM_USER_DATA = "USER.START_GET_PARENT_MOIM_USER_DATA",
  SUCCEEDED_GET_PARENT_MOIM_USER_DATA = "USER.SUCCEEDED_GET_PARENT_MOIM_USER_DATA",
  FAILED_GET_PARENT_MOIM_USER_DATA = "USER.FAILED_GET_PARENT_MOIM_USER_DATA",

  START_GET_PROFILE_BLOCKS = "USER.START_GET_PROFILE_BLOCKS",
  SUCCEEDED_GET_PROFILE_BLOCKS = "USER.SUCCEEDED_GET_PROFILE_BLOCKS",
  FAILED_GET_PROFILE_BLOCKS = "USER.FAILED_GET_PROFILE_BLOCKS",

  START_SEARCH_PAGE_SEARCH_USERS = "USER.START_SEARCH_PAGE_SEARCH_USERS",
  SUCCEED_SEARCH_PAGE_SEARCH_USERS = "USER.SUCCEED_SEARCH_PAGE_SEARCH_USERS",
  FAILED_SEARCH_PAGE_SEARCH_USERS = "USER.FAILED_SEARCH_PAGE_SEARCH_USERS",
  CLEAR_SEARCH_PAGE_SEARCH_USERS = "USER.CLEAR_SEARCH_PAGE_SEARCH_USERS",
}
// ...
export enum MeTypes {
  OPEN_EDIT_MY_PROFILE = "ME.OPEN_EDIT_MY_PROFILE",
  CLOSE_EDIT_MY_PROFILE = "ME.CLOSE_EDIT_MY_PROFILE",

  START_CHANGE_MY_PROFILE = "ME.START_CHANGE_MY_PROFILE",
  SUCCEEDED_CHANGE_MY_PROFILE = "ME.SUCCEEDED_CHANGE_MY_PROFILE",
  FAILED_CHANGE_MY_PROFILE = "ME.FAILED_CHANGE_MY_PROFILE",

  START_UPDATE_AVATAR = "ME.START_UPDATE_AVATAR",
  SUCCEEDED_UPDATE_AVATAR = "ME.SUCCEEDED_UPDATE_AVATAR",
  FAILED_UPDATE_AVATAR = "ME.FAILED_UPDATE_AVATAR",

  START_GET_MY_JOINED_MOIMS = "ME.START_GET_MY_JOINED_MOIMS",
  SUCCEEDED_GET_MY_JOINED_MOIMS = "ME.SUCCEEDED_GET_MY_JOINED_MOIMS",
  FAILED_GET_MY_JOINED_MOIMS = "ME.FAILED_GET_MY_JOINED_MOIMS",

  ADD_MY_JOINED_MOIM = "ME.ADD_MY_JOINED_MOIM",

  START_LEAVE_MOIM = "ME.START_LEAVE_MOIM",
  SUCCEED_LEAVE_MOIM = "ME.SUCCEED_LEAVE_MOIM",
  FAILED_LEAVE_MOIM = "ME.FAILED_LEAVE_MOIM",

  OPEN_FAILED_LEAVE_MOIM_ALERT = "ME.OPEN_FAILED_LEAVE_MOIM_ALERT",
  CLOSE_FAILED_LEAVE_MOIM_ALERT = "ME.CLOSE_FAILED_LEAVE_MOIM_ALERT",

  START_GET_BOOKMARKS = "ME.START_GET_BOOKMARKS",
  SUCCEEDED_GET_BOOKMARKS = "ME.SUCCEEDED_GET_BOOKMARKS",
  FAILED_GET_BOOKMARKS = "ME.FAILED_GET_BOOKMARKS",
  START_POST_BOOKMARK = "ME.START_POST_BOOKMARK",
  SUCCEEDED_POST_BOOKMARK = "ME.SUCCEEDED_POST_BOOKMARK",
  FAILED_POST_BOOKMARK = "ME.FAILED_POST_BOOKMARK",
  START_DELETE_BOOKMARK = "ME.START_DELETE_BOOKMARK",
  SUCCEEDED_DELETE_BOOKMARK = "ME.SUCCEEDED_DELETE_BOOKMARK",
  FAILED_DELETE_BOOKMARK = "ME.FAILED_DELETE_BOOKMARK",
  START_ALL_MOIM_RESOURCES = "ME.START_ALL_MOIM_RESOURCES",
  SUCCEEDED_ALL_MOIM_RESOURCES = "ME.SUCCEEDED_ALL_MOIM_RESOURCES",
  FAILED_ALL_MOIM_RESOURCES = "ME.FAILED_ALL_MOIM_RESOURCES",

  CHANGE_BOOKMAKR_CURRENT_MOIM_RESOURCE = "ME.CHANGE_BOOKMAKR_CURRENT_MOIM_RESOURCE",
}